Move LowlevelException to Bolt\Exception
