+++ tests/Configuration/LowlevelChecksTest.php
@@ -3,9 +3,9 @@ namespace Bolt\Tests\Configuration;
 
 use Bolt\Configuration\Standard;
 use Bolt\Configuration\LowlevelChecks;
-use Bolt\Configuration\LowlevelException;
-use Bolt\Tests\BoltUnitTest;
 use Bolt\Configuration\ResourceManager;
+use Bolt\Exception\LowlevelException;
+use Bolt\Tests\BoltUnitTest;
 
 /**
  * Class to test src/Configuration/LowlevelChecks.php.
@@ -24,7 +24,7 @@ class LowlevelChecksTest extends BoltUnitTest
 
     public function setUp()
     {
-        $this->php = \PHPUnit_Extension_FunctionMocker::start($this, 'Bolt\Configuration')
+        $this->php = \PHPUnit_Extension_FunctionMocker::start($this, 'Bolt\Exception')
             ->mockFunction('is_readable')
             ->mockFunction('is_writable')
             ->mockFunction('file_exists')
@@ -76,7 +76,7 @@ class LowlevelChecksTest extends BoltUnitTest
     {
         $check = $this->getCleanChecker();
         $check->magicQuotes = true;
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt - Fatal Error/");
         $check->doChecks();
     }
@@ -85,7 +85,7 @@ class LowlevelChecksTest extends BoltUnitTest
     {
         $check = $this->getCleanChecker();
         $check->safeMode = true;
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt - Fatal Error/");
         $check->doChecks();
     }
@@ -94,7 +94,7 @@ class LowlevelChecksTest extends BoltUnitTest
     {
         $check = $this->getCleanChecker();
         $check->isApache = true;
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt - Fatal Error/");
         $check->doChecks();
     }
@@ -111,7 +111,7 @@ class LowlevelChecksTest extends BoltUnitTest
     {
         $check = $this->getMockedChecker('mockMysql');
         $check->mysqlLoaded = false;
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt - Fatal Error/");
         $check->doDatabaseCheck();
     }
@@ -141,7 +141,7 @@ class LowlevelChecksTest extends BoltUnitTest
     {
         $check = $this->getMockedChecker('mockPostgres');
         $check->postgresLoaded = false;
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt - Fatal Error/");
         $check->doDatabaseCheck();
     }
@@ -150,7 +150,7 @@ class LowlevelChecksTest extends BoltUnitTest
     {
         $check = $this->getMockedChecker('mockSqlite');
         $check->sqliteLoaded = false;
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt - Fatal Error/");
         $check->doDatabaseCheck();
     }
@@ -158,7 +158,7 @@ class LowlevelChecksTest extends BoltUnitTest
     public function testPlatformUnsupported()
     {
         $check = $this->getMockedChecker('mockUnsupportedPlatform');
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/database type, but it is not supported/");
         $check->doDatabaseCheck();
     }
@@ -218,7 +218,7 @@ class LowlevelChecksTest extends BoltUnitTest
             ->with('test/bolt.db')
             ->will($this->returnValue(false));
 
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/is not writable/");
         $check->doDatabaseCheck();
     }
@@ -238,7 +238,7 @@ class LowlevelChecksTest extends BoltUnitTest
             ->with('test')
             ->will($this->returnValue(false));
 
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/does not exist/");
         $check->doDatabaseCheck();
     }
@@ -264,7 +264,7 @@ class LowlevelChecksTest extends BoltUnitTest
             ->with('test')
             ->will($this->returnValue(false));
 
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/is not writable/");
         $check->doDatabaseCheck();
     }
@@ -272,7 +272,7 @@ class LowlevelChecksTest extends BoltUnitTest
     public function testDbFailsAsRootWithoutPassword()
     {
         $check = $this->getMockedChecker('mockRoot');
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt will stubbornly refuse to run until you've set a password/");
         $check->doDatabaseCheck();
     }
@@ -280,7 +280,7 @@ class LowlevelChecksTest extends BoltUnitTest
     public function testEmptyDb()
     {
         $check = $this->getMockedChecker('mockEmptyDb');
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt - Fatal Error/");
         $check->doDatabaseCheck();
     }
@@ -288,7 +288,7 @@ class LowlevelChecksTest extends BoltUnitTest
     public function testEmptyDbUser()
     {
         $check = $this->getMockedChecker('mockEmptyUser');
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt - Fatal Error/");
         $check->doDatabaseCheck();
     }
@@ -353,7 +353,7 @@ class LowlevelChecksTest extends BoltUnitTest
     {
         $badDir = "/path/to/nowhere";
         $check = $this->getCleanChecker();
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt - Fatal Error/");
         $check->assertWritableDir($badDir);
     }
@@ -400,7 +400,7 @@ class LowlevelChecksTest extends BoltUnitTest
             ->method('is_readable')
             ->will($this->returnValue(false));
 
-        $this->setExpectedException('Bolt\Configuration\LowlevelException');
+        $this->setExpectedException('Bolt\Exception\LowlevelException');
         $this->expectOutputRegex("/Bolt - Fatal Error/");
         $check->doChecks();
     }
